+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - The kernel was not completely resetted
+- The client is no longer kept in memory as it conflicts when the kernel is shutdown
 
 ## 1.1.0 - 2023-09-02